test: 重构测试用例按功能分块
