+++ test/mock/localstorage.ts
@@ -1,0 +1,40 @@
+class LocalStorageMock {
+  store: any
+
+  get length() {
+    return Object.keys(this.store).length
+  }
+
+  key(index: number) {
+    const keys = Object.keys(this.store)
+    if (index >= keys.length) return null
+
+    return keys[index]
+  }
+
+  constructor() {
+    this.store = {}
+  }
+
+  clear() {
+    this.store = {}
+  }
+
+  getItem(key: string) {
+    return this.store[key] || null
+  }
+
+  setItem(key: string, value: string) {
+    this.store[key] = String(value)
+  }
+
+  removeItem(key: string) {
+    delete this.store[key]
+  }
+
+  clean() {
+    this.store = {}
+  }
+}
+
+export default LocalStorageMock
